requirements.txt
Flask==2.3.2
pypdf==3.9.0
python-dotenv==1.0.0
anthropic>=0.7.0